+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Add solution method to solution of vt object. [#166](https://github.com/greenbone/ospd/pull/166)
 
+### Changes
+- Modify __init__() method and use new syntax for super(). [#186](https://github.com/greenbone/ospd/pull/186)
+
 ## [2.0.1] (unreleased)
 
 ### Added